backend_api/requirements.txt
fastapi==0.110.0
pydantic==2.5.2
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
python-dotenv==1.0.0
lxml==4.9.3
openai==1.6.1
starlette==0.36.3
tenacity==8.2.3
httpx==0.27.0
